#7 xml namespace, pom, and import fixes
