+++ roaster-api/src/main/java/hu/icellmobilsoft/roaster/api/InvalidConfigException.java
@@ -1,0 +1,67 @@
+/*-
+ * #%L
+ * Coffee
+ * %%
+ * Copyright (C) 2020 i-Cell Mobilsoft Zrt.
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package hu.icellmobilsoft.roaster.api;
+
+import javax.enterprise.inject.Vetoed;
+
+/**
+ * Exception class indicating errors regarding the Roaster configuration.
+ *
+ * @author martin.nagy
+ * @since 0.2.0
+ */
+@Vetoed
+public class InvalidConfigException extends TestException {
+
+    /**
+     * @see TestException#TestException(String)
+     * @param  message the detail message (which is saved for later retrieval
+     *         by the {@link #getMessage()} method).
+     */
+    public InvalidConfigException(String message) {
+        super(message);
+    }
+
+
+    /**
+     * @see TestException#TestException(String, Throwable)
+     * @param  message the detail message (which is saved for later retrieval
+     *         by the {@link #getMessage()} method).
+     * @param  cause the cause (which is saved for later retrieval by the
+     *         {@link #getCause()} method).  (A {@code null} value is
+     *         permitted, and indicates that the cause is nonexistent or
+     *         unknown.)
+     */
+    public InvalidConfigException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+
+    /**
+     * @see TestException#TestException(Throwable)
+     * @param  cause the cause (which is saved for later retrieval by the
+     *         {@link #getCause()} method).  (A {@code null} value is
+     *         permitted, and indicates that the cause is nonexistent or
+     *         unknown.)
+     */
+    public InvalidConfigException(Throwable cause) {
+        super(cause);
+    }
+}
