+++ roaster-tm4j/roaster-tm4j-common/src/main/java/hu/icellmobilsoft/roaster/tm4j/common/Tm4jRest.java
@@ -38,5 +38,5 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Target({ TYPE, METHOD, PARAMETER, FIELD })
 @Retention(RUNTIME)
 @Qualifier
-public @interface Rest {
+public @interface Tm4jRest {
 }
