+++ roaster-tm4j/roaster-tm4j-dto/src/main/resources/xsd/hu/icellmobilsoft/roaster/tm4j/dto/domain/test-execution.xsd
@@ -1,6 +1,6 @@
-<xs:schema xmlns="http://tm4j.dto.roaster.icellmobilsoft.hu/test-execution"
+<xs:schema xmlns="http://domain.dto.tm4j.roaster.icellmobilsoft.hu/test-execution"
            xmlns:xs="http://www.w3.org/2001/XMLSchema"
-           targetNamespace="http://tm4j.dto.roaster.icellmobilsoft.hu/test-execution"
+           targetNamespace="http://domain.dto.tm4j.roaster.icellmobilsoft.hu/test-execution"
            elementFormDefault="qualified">
 
     <xs:element name="Execution">
