+++ roaster-tm4j/roaster-tm4j-dto/src/main/resources/xsd/hu/icellmobilsoft/roaster/tm4j/dto/super.catalog.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
 <catalog xmlns="urn:oasis:names:tc:entity:xmlns:xml:catalog">
-    <public publicId="http://tm4j.dto.roaster.icellmobilsoft.hu/test-execution" uri="./domain/test-execution.xsd" />
+    <public publicId="http://domain.dto.tm4j.roaster.icellmobilsoft.hu/test-execution" uri="./domain/test-execution.xsd" />
 </catalog>
