+++ roaster-tm4j/roaster-tm4j-dto/src/main/resources/xsd/hu/icellmobilsoft/roaster/tm4j/dto/super.xsd
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <xsd:schema xmlns:jaxb="http://java.sun.com/xml/ns/jaxb"
     xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-    xmlns="http://super.dto.icon.icellmobilsoft.hu/super"
+    xmlns="http://super.dto.tm4j.roaster.icellmobilsoft.hu/super"
     xmlns:xjc="http://java.sun.com/xml/ns/jaxb/xjc" 
-    targetNamespace="http://super.dto.icon.icellmobilsoft.hu/super"
+    targetNamespace="http://super.dto.tm4j.roaster.icellmobilsoft.hu/super"
     elementFormDefault="qualified"
     jaxb:version="2.0"
     jaxb:extensionBindingPrefixes="xjc">
 
-    <xsd:import namespace="http://tm4j.dto.roaster.icellmobilsoft.hu/test-execution" />
+    <xsd:import namespace="http://domain.dto.tm4j.roaster.icellmobilsoft.hu/test-execution" />
 </xsd:schema>
